+++ argilla-frontend/v1/domain/usecases/dataset-setting/get-dataset-settings-use-case.ts
@@ -88,7 +88,6 @@ export class GetDatasetSettingsUseCase {
         field.id,
         field.name,
         field.title,
-        "",
         datasetId,
         field.required,
         field.settings